Refactor Cody Web Chat wrapper
