+++ web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.7.0
+- Simplifies Cody Web Chat component (it requires now using only one root CodyWebChat component)
+- Adds skeleton loading state to Cody Web Chat UI 
+
 ## 0.6.1
 - Adds support for directory initial context (mention) 
 - Changes target build to es modules
