+++ web/lib/components/CodyWebChat.module.css
@@ -1,0 +1,13 @@
+
+.root {
+    display: flex;
+    height: 100vh;
+    width: 100%;
+    font-size: 13px;
+    overflow: hidden;
+}
+
+.container {
+    height: 100%;
+    flex-grow: 1;
+}
