+++ web/lib/components/CodyWebChat.tsx
@@ -1,4 +1,5 @@
-import { type FC, useCallback, useLayoutEffect, useMemo, useState } from 'react'
+import classNames from 'classnames'
+import { type FC, type FunctionComponent, useCallback, useLayoutEffect, useMemo, useState } from 'react'
 import { URI } from 'vscode-uri'
 
 import {
@@ -16,27 +17,28 @@ import {
     isErrorLike,
     setDisplayPathEnvInfo,
 } from '@sourcegraph/cody-shared'
+import { AppWrapper } from 'cody-ai/webviews/AppWrapper'
+import type { VSCodeWrapper } from 'cody-ai/webviews/utils/VSCodeApi'
 
 import { ChatMentionContext, type ChatMentionsSettings } from '@sourcegraph/prompt-editor'
 import { getAppWrappers } from 'cody-ai/webviews/App'
+import { CodyPanel } from 'cody-ai/webviews/CodyPanel'
 import { ChatEnvironmentContext } from 'cody-ai/webviews/chat/ChatEnvironmentContext'
 import type { ChatModelContext } from 'cody-ai/webviews/chat/models/chatModelContext'
 import { useClientActionDispatcher } from 'cody-ai/webviews/client/clientState'
+import type { View } from 'cody-ai/webviews/tabs'
+import { ComposedWrappers, type Wrapper } from 'cody-ai/webviews/utils/composeWrappers'
 import { createWebviewTelemetryRecorder } from 'cody-ai/webviews/utils/telemetry'
+import type { Config } from 'cody-ai/webviews/utils/useConfig'
 
-import { useWebAgentClient } from './CodyWebPanelProvider'
+import type { InitialContext } from '../types'
+
+import { useCodyWebAgent } from './use-cody-agent'
 
 // Include global Cody Web styles to the styles bundle
 import '../global-styles/styles.css'
-import { CodyPanel } from 'cody-ai/webviews/CodyPanel'
-import type { View } from 'cody-ai/webviews/tabs'
-import { ComposedWrappers, type Wrapper } from 'cody-ai/webviews/utils/composeWrappers'
-import type { Config } from 'cody-ai/webviews/utils/useConfig'
-import styles from './CodyWebPanel.module.css'
-
-const CONTEXT_MENTIONS_SETTINGS: ChatMentionsSettings = {
-    resolutionMode: 'remote',
-}
+import styles from './CodyWebChat.module.css'
+import { ChatSkeleton } from './skeleton/ChatSkeleton'
 
 // Internal API mock call in order to set up web version of
 // the cody agent properly (completely mock data)
@@ -45,18 +47,72 @@ setDisplayPathEnvInfo({
     workspaceFolders: [URI.file('/tmp/foo')],
 })
 
-export interface CodyWebPanelProps {
+export interface CodyWebChatProps {
+    serverEndpoint: string
+    accessToken: string | null
+    createAgentWorker: () => Worker
+    telemetryClientName?: string
+    initialContext?: InitialContext
+    customHeaders?: Record<string, string>
+    className?: string
+}
+/**
+ * The root store/provider node for Cody Web, creates and shares
+ * agent client and maintains active web panel ID, chat history and vscodeAPI.
+ */
+export const CodyWebChat: FunctionComponent<CodyWebChatProps> = ({
+    serverEndpoint,
+    accessToken,
+    createAgentWorker,
+    initialContext,
+    telemetryClientName,
+    customHeaders,
+    className,
+}) => {
+    const { client, vscodeAPI } = useCodyWebAgent({
+        serverEndpoint,
+        accessToken,
+        createAgentWorker,
+        initialContext,
+        telemetryClientName,
+        customHeaders,
+    })
+
+    if (isErrorLike(client)) {
+        return <p>Cody Web client agent error: {client.message}</p>
+    }
+
+    if (client === null || vscodeAPI === null) {
+        return <ChatSkeleton className={classNames(className, styles.root)} />
+    }
+
+    return (
+        <AppWrapper>
+            <div className={classNames(className, styles.root)}>
+                <CodyWebPanel
+                    vscodeAPI={vscodeAPI}
+                    initialContext={initialContext}
+                    className={styles.container}
+                />
+            </div>
+        </AppWrapper>
+    )
+}
+
+const CONTEXT_MENTIONS_SETTINGS: ChatMentionsSettings = {
+    resolutionMode: 'remote',
+}
+
+interface CodyWebPanelProps {
+    vscodeAPI: VSCodeWrapper
+    initialContext: InitialContext | undefined
     className?: string
 }
 
-// NOTE: This code is implements a subset of the
-// functionality for the experimental web chat prototype.
-export const CodyWebPanel: FC<CodyWebPanelProps> = props => {
-    const { className } = props
+const CodyWebPanel: FC<CodyWebPanelProps> = props => {
+    const { vscodeAPI, initialContext, className } = props
 
-    const { vscodeAPI, client, initialContext } = useWebAgentClient()
     const dispatchClientAction = useClientActionDispatcher()
-
     const [errorMessages, setErrorMessages] = useState<string[]>([])
     const [isTranscriptError, setIsTranscriptError] = useState<boolean>(false)
     const [messageInProgress, setMessageInProgress] = useState<ChatMessage | null>(null)
@@ -213,39 +269,33 @@ export const CodyWebPanel: FC<CodyWebPanelProps> = props => {
         [vscodeAPI, telemetryRecorder, chatModelContext, clientState, config, envVars]
     )
 
-    const isLoading = !client || !chatModels || !config || !view || !userHistory
+    const isLoading = !chatModels || !config || !view || !userHistory
 
     return (
         <div className={className} data-cody-web-chat={true}>
-            {!isLoading ? (
-                isErrorLike(client) ? (
-                    <p>Error: {client.message}</p>
-                ) : (
-                    <ChatEnvironmentContext.Provider value={envVars}>
-                        <ChatMentionContext.Provider value={CONTEXT_MENTIONS_SETTINGS}>
-                            <ComposedWrappers wrappers={wrappers}>
-                                <CodyPanel
-                                    view={view}
-                                    setView={setView}
-                                    errorMessages={errorMessages}
-                                    setErrorMessages={setErrorMessages}
-                                    attributionEnabled={false}
-                                    config={config.config}
-                                    userHistory={userHistory}
-                                    chatEnabled={true}
-                                    showWelcomeMessage={true}
-                                    showIDESnippetActions={false}
-                                    messageInProgress={messageInProgress}
-                                    transcript={transcript}
-                                    vscodeAPI={vscodeAPI}
-                                    isTranscriptError={isTranscriptError}
-                                />
-                            </ComposedWrappers>
-                        </ChatMentionContext.Provider>
-                    </ChatEnvironmentContext.Provider>
-                )
-            ) : (
-                <div className={styles.loading}>Loading Cody Client...</div>
+            {!isLoading && (
+                <ChatEnvironmentContext.Provider value={envVars}>
+                    <ChatMentionContext.Provider value={CONTEXT_MENTIONS_SETTINGS}>
+                        <ComposedWrappers wrappers={wrappers}>
+                            <CodyPanel
+                                view={view}
+                                setView={setView}
+                                errorMessages={errorMessages}
+                                setErrorMessages={setErrorMessages}
+                                attributionEnabled={false}
+                                config={config.config}
+                                userHistory={userHistory}
+                                chatEnabled={true}
+                                showWelcomeMessage={true}
+                                showIDESnippetActions={false}
+                                messageInProgress={messageInProgress}
+                                transcript={transcript}
+                                vscodeAPI={vscodeAPI}
+                                isTranscriptError={isTranscriptError}
+                            />
+                        </ComposedWrappers>
+                    </ChatMentionContext.Provider>
+                </ChatEnvironmentContext.Provider>
             )}
         </div>
     )
