--- web/lib/components/CodyWebPanel.module.css
@@ -1,7 +1,0 @@
-.loading {
-    margin: auto;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-}
