+++ web/lib/index.ts
@@ -1,4 +1,4 @@
-export { CodyWebPanelProvider } from './components/CodyWebPanelProvider'
-export { CodyWebPanel, type CodyWebPanelProps } from './components/CodyWebPanel'
+export { CodyWebChat, type CodyWebChatProps } from './components/CodyWebChat'
+export { ChatSkeleton } from './components/skeleton/ChatSkeleton'
 
 export type { Repository, InitialContext } from './types'
